+++ .storybook/preview.js
@@ -1,5 +1,5 @@
 /* eslint-env browser */
-import { addDecorator } from '@storybook/react';
+import { addDecorator, addParameters } from '@storybook/react';
 import ansiHTML from 'ansi-html';
 import chalk from 'chalk';
 import React from 'react';
@@ -28,6 +28,10 @@ const style = {
   lineHeight: '1rem',
   color: '#c0c4cd',
 };
+
+addParameters({
+  layout: 'fullscreen',
+});
 
 addDecorator((storyFn, { kind }) => {
   if (kind.startsWith('CLI/')) {