+++ action/main.js
@@ -55,6 +55,7 @@ var github_1 = require("@actions/github");
 var jsonfile_1 = require("jsonfile");
 var pkg_up_1 = __importDefault(require("pkg-up"));
 var uuid_1 = require("uuid");
+var path_1 = __importDefault(require("path"));
 var getEnv_1 = __importDefault(require("../bin/lib/getEnv"));
 var log_1 = require("../bin/lib/log");
 var parseArgs_1 = __importDefault(require("../bin/lib/parseArgs"));
@@ -73,7 +74,9 @@ var maybe = function (a, b) {
 };
 var getCommit = function (event) {
     switch (event.eventName) {
-        case 'pull_request': {
+        case 'pull_request':
+        case 'pull_request_review':
+        case 'pull_request_target': {
             return {
                 // @ts-ignore
                 owner: event.payload.repository.owner.login,
@@ -138,7 +141,7 @@ function runChromatic(options) {
 }
 function run() {
     return __awaiter(this, void 0, void 0, function () {
-        var commit, branch, sha, projectToken, buildScriptName, scriptName, exec, skip, doNotStart, storybookPort, storybookUrl, storybookBuildDir, storybookHttps, storybookCert, storybookKey, storybookCa, preserveMissing, autoAcceptChanges, allowConsoleErrors, exitZeroOnChanges, exitOnceUploaded, ignoreLastBuildOnBranch, chromatic, _a, url, code, e_1;
+        var commit, branch, sha, projectToken, workingDir, buildScriptName, scriptName, exec, skip, doNotStart, storybookPort, storybookUrl, storybookBuildDir, storybookHttps, storybookCert, storybookKey, storybookCa, preserveMissing, autoAcceptChanges, allowConsoleErrors, exitZeroOnChanges, exitOnceUploaded, ignoreLastBuildOnBranch, chromatic, _a, url, code, e_1;
         return __generator(this, function (_b) {
             switch (_b.label) {
                 case 0:
@@ -151,6 +154,7 @@ function run() {
                 case 1:
                     _b.trys.push([1, 3, , 4]);
                     projectToken = core_1.getInput('projectToken') || core_1.getInput('appCode');
+                    workingDir = core_1.getInput('workingDir');
                     buildScriptName = core_1.getInput('buildScriptName');
                     scriptName = core_1.getInput('scriptName');
                     exec = core_1.getInput('exec');
@@ -171,8 +175,10 @@ function run() {
                     ignoreLastBuildOnBranch = core_1.getInput('ignoreLastBuildOnBranch');
                     process.env.CHROMATIC_SHA = sha;
                     process.env.CHROMATIC_BRANCH = branch;
+                    process.chdir(path_1["default"].join(process.cwd(), workingDir || ''));
                     chromatic = runChromatic({
                         projectToken: projectToken,
+                        workingDir: maybe(workingDir),
                         buildScriptName: maybe(buildScriptName),
                         scriptName: maybe(scriptName),
                         exec: maybe(exec),