+++ bin/tasks/gitInfo.js
@@ -26,7 +26,7 @@ export const setGitInfo = async (ctx, task) => {
   ctx.git.version = await getVersion();
   const { branch, commit } = ctx.git;
 
-  const matchesBranch = glob => (glob && glob.length ? picomatch(glob)(branch) : !!glob);
+  const matchesBranch = (glob) => (glob && glob.length ? picomatch(glob)(branch) : !!glob);
   ctx.git.matchesBranch = matchesBranch;
 
   if (matchesBranch(skip)) {