+++ bin/ui/messages/info/wroteReport.js
@@ -2,4 +2,4 @@ import chalk from 'chalk';
 
 import { info } from '../../components/icons';
 
-export default filePath => chalk`${info} Wrote JUnit XML report to {bold ${filePath}}`;
+export default (filePath) => chalk`${info} Wrote JUnit XML report to {bold ${filePath}}`;