+++ bin/ui/tasks/report.js
@@ -3,13 +3,13 @@ export const initial = {
   title: 'Generate build report',
 };
 
-export const pending = ctx => ({
+export const pending = (ctx) => ({
   status: 'pending',
   title: 'Generating build report',
   output: `Collecting build information`,
 });
 
-export const success = ctx => ({
+export const success = (ctx) => ({
   status: 'success',
   title: `Generated build report`,
   output: `View report at ${ctx.reportPath}`,