+++ bin/ui/tasks/storybookInfo.js
@@ -1,13 +1,13 @@
-const capitalize = string =>
+const capitalize = (string) =>
   string
     .split('-')
-    .map(str => str.charAt(0).toUpperCase() + str.slice(1))
+    .map((str) => str.charAt(0).toUpperCase() + str.slice(1))
     .join(' ');
 
 const infoMessage = ({ addons, version, viewLayer }) => {
   const info = `Storybook v${version} for ${capitalize(viewLayer)}`;
   return addons.length
-    ? `${info}; supported addons found: ${addons.map(addon => capitalize(addon.name)).join(', ')}`
+    ? `${info}; supported addons found: ${addons.map((addon) => capitalize(addon.name)).join(', ')}`
     : `${info}; no supported addons found`;
 };
 
@@ -16,12 +16,12 @@ export const initial = {
   title: 'Collect Storybook metadata',
 };
 
-export const pending = ctx => ({
+export const pending = (ctx) => ({
   status: 'pending',
   title: 'Collecting Storybook metadata',
 });
 
-export const success = ctx => ({
+export const success = (ctx) => ({
   status: 'success',
   title: 'Collected Storybook metadata',
   output: infoMessage(ctx.storybook),