+++ bin/ui/workflows/tunnelBuild.stories.js
@@ -9,11 +9,11 @@ import * as storybookInfo from '../tasks/storybookInfo.stories';
 import * as tunnel from '../tasks/tunnel.stories';
 import * as verify from '../tasks/verify.stories';
 
-const steps = (...arr) => arr.map(step => task(step())).join('\n');
+const steps = (...arr) => arr.map((step) => task(step())).join('\n');
 
 export default {
   title: 'CLI/Workflows/TunnelBuild',
-  decorators: [storyFn => storyFn().join('\n\n')],
+  decorators: [(storyFn) => storyFn().join('\n\n')],
 };
 
 export const Initial = () => [